requirements.txt
Django
SQLAlchemy
exrex
fake-factory
gunicorn
psycopg2
wsgiref
python-social-auth